Extract test driver factory and make deleteFile suspend
Pre-web-parity refactor: introduce expect testSqliteDriver() in
:integration and :coroutines-extensions test scaffolding, and convert
the deleteFile expect/actual to a suspend function so a webTest actual
can use the OPFS removeEntry suspend API later. withDatabase helpers in
:library tests are also marked suspend to propagate.
