+++ coroutines-extensions/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/coroutines/CommonFlowExtensionsTests.kt
@@ -1,7 +1,3 @@
 package com.eygraber.sqldelight.androidx.driver.coroutines
 
-import androidx.sqlite.SQLiteDriver
-
 expect class CommonFlowExtensionsTest() : FlowExtensionsTest
-
-expect fun testSQLiteDriver(): SQLiteDriver
