+++ coroutines-extensions/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/coroutines/TestSqliteDriver.kt
@@ -1,0 +1,5 @@
+package com.eygraber.sqldelight.androidx.driver.coroutines
+
+import androidx.sqlite.SQLiteDriver
+
+expect fun testSqliteDriver(): SQLiteDriver
