+++ integration/src/jvmTest/kotlin/com/eygraber/sqldelight/androidx/driver/integration/AndroidxSqliteCommonIntegrationTests.jvm.kt
@@ -2,6 +2,6 @@ package com.eygraber.sqldelight.androidx.driver.integration
 
 import java.io.File
 
-actual fun deleteFile(name: String) {
+actual suspend fun deleteFile(name: String) {
   File(name).delete()
 }
