+++ integration/src/nativeTest/kotlin/com/eygraber/sqldelight/androidx/driver/integration/AndroidxSqliteCommonIntegrationTests.native.kt
@@ -3,6 +3,6 @@ package com.eygraber.sqldelight.androidx.driver.integration
 import okio.FileSystem
 import okio.Path.Companion.toPath
 
-actual fun deleteFile(name: String) {
+actual suspend fun deleteFile(name: String) {
   FileSystem.SYSTEM.delete(name.toPath())
 }
