+++ integration/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/integration/AndroidxSqliteIntegrationTest.kt
@@ -1,10 +1,10 @@
 package com.eygraber.sqldelight.androidx.driver.integration
 
-import androidx.sqlite.driver.bundled.BundledSQLiteDriver
 import app.cash.sqldelight.SuspendingTransactionWithoutReturn
 import com.eygraber.sqldelight.androidx.driver.AndroidxSqliteConfiguration
 import com.eygraber.sqldelight.androidx.driver.AndroidxSqliteDatabaseType
 import com.eygraber.sqldelight.androidx.driver.AndroidxSqliteDriver
+import kotlinx.coroutines.test.runTest
 import kotlin.test.AfterTest
 
 abstract class AndroidxSqliteIntegrationTest {
@@ -22,7 +22,7 @@ abstract class AndroidxSqliteIntegrationTest {
 
   val driver by lazy {
     AndroidxSqliteDriver(
-      driver = BundledSQLiteDriver(),
+      driver = testSqliteDriver(),
       databaseType = type,
       schema = AndroidXDb.Schema,
       configuration = createConfiguration(),
@@ -34,7 +34,7 @@ abstract class AndroidxSqliteIntegrationTest {
   }
 
   @AfterTest
-  fun cleanup() {
+  fun cleanup() = runTest {
     driver.close()
 
     (type as? AndroidxSqliteDatabaseType.File)?.let { type ->
