+++ integration/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/integration/TestSqliteDriver.kt
@@ -1,0 +1,5 @@
+package com.eygraber.sqldelight.androidx.driver.integration
+
+import androidx.sqlite.SQLiteDriver
+
+expect fun testSqliteDriver(): SQLiteDriver
