+++ library/src/jvmTest/kotlin/com/eygraber/sqldelight/androidx/driver/AndroidxSqliteCommonTests.jvm.kt
@@ -25,7 +25,7 @@ actual fun androidxSqliteTestDriver(): SQLiteDriver = BundledSQLiteDriver()
 @Suppress("InjectDispatcher")
 actual val IoDispatcher: CoroutineDispatcher get() = Dispatchers.IO
 
-actual fun deleteFile(name: String) {
+actual suspend fun deleteFile(name: String) {
   File(name).delete()
 }
