+++ library/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/AndroidxSqliteCallbackTest.kt
@@ -99,7 +99,7 @@ abstract class AndroidxSqliteCallbackTest {
     }
   }
 
-  private inline fun withDatabase(
+  private suspend inline fun withDatabase(
     schema: SqlSchema<QueryResult.AsyncValue<Unit>>,
     dbName: String,
     noinline onConfigure: suspend AndroidxSqliteConfigurableDriver.() -> Unit,
