+++ library/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/AndroidxSqliteConcurrencyTest.kt
@@ -89,7 +89,7 @@ abstract class AndroidxSqliteConcurrencyTest {
     }
   }
 
-  private inline fun withDatabase(
+  private suspend inline fun withDatabase(
     schema: SqlSchema<QueryResult.AsyncValue<Unit>>,
     dbName: String,
     noinline onCreate: suspend SqlDriver.() -> Unit,
