+++ library/src/nonWebTest/kotlin/com/eygraber/sqldelight/androidx/driver/AndroidxSqliteMigrationTest.kt
@@ -104,7 +104,7 @@ abstract class AndroidxSqliteMigrationTest {
     }
   }
 
-  private inline fun withDatabase(
+  private suspend inline fun withDatabase(
     schema: SqlSchema<QueryResult.AsyncValue<Unit>>,
     dbName: String,
     noinline onCreate: suspend SqlDriver.() -> Unit,
